+++ apps/web/src/components/button/sumit.tsx
@@ -1,21 +1,5 @@
 import tw, { styled } from 'twin.macro';
 
-import ClosingSvg from '@/assests/icon/icon-closing.svg';
-
-interface CloseButtonProps {
-  onClose: () => void;
-}
-export default function CloseButton({ onClose }: CloseButtonProps) {
-  return (
-    <button
-      className="absolute top-3 right-3 z-10"
-      onClick={onClose}
-      type="button"
-    >
-      <ClosingSvg />
-    </button>
-  );
-}
 interface SubmitButtonProps {
   isActive: boolean;
 }
